% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo_stream_in_dt.R
\name{mongo_stream_in_dt}
\alias{mongo_stream_in_dt}
\title{Read in MongoDB data}
\usage{
mongo_stream_in_dt(cur, pagesize = 1000, verbose = TRUE)
}
\arguments{
\item{cur}{MongoDB Cursor}

\item{pagesize}{Pagesize}

\item{verbose}{Verbose}
}
\description{
Takes a mongodb cursor and reads the data into a data.table object.
}

