% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_rbind_page.R
\name{try_rbind_page}
\alias{try_rbind_page}
\title{try_rbind_page}
\usage{
try_rbind_page(lst)
}
\arguments{
\item{lst}{List object of the returned MongoDB data}
}
\description{
attempt to data.table::rbindlist the page returned by the mongo cursor
}

