% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_mongolitedt.R
\name{bind_mongolitedt}
\alias{bind_mongolitedt}
\title{Bind_mongolitedt}
\usage{
bind_mongolitedt(mongo)
}
\arguments{
\item{mongo}{A mongolite::mongo connection object to which the functions are bound}
}
\value{
mongo The mongo object with additional functions
}
\description{
Bind functions to a mongolite::mongo object
}

